refactor(auth): Restructure login form components for improved maintainability
- Moved the login form logic into separate components: `AnonymousLoginForm` and `DefaultLoginForm`, enhancing modularity and readability.
- Updated the main `LoginForm` component to conditionally render the appropriate form based on the demo mode.
- Removed the old `login-form.tsx` file to streamline the codebase and reduce complexity.
