+++ app/(auth-pages)/login/page.tsx
@@ -1,4 +1,4 @@
-import { LoginForm } from "@/components/login-form";
+import { LoginForm } from "@/components/login-form/form";
 
 export default function Login() {
   return (
